Checking in stable vim configuration.
Truth is, I've been using this as it is for a long time.

But I'm about to do the big update to neovim 0.5 and I know there's going to be some churn as I try
to simplify the config as much as possible. This commit provides a solid snapshot of changes I can
refer back to as needed.